Re-integrate MemImpl into MemModel
