+++ crucible-llvm/crucible-llvm.cabal
@@ -57,7 +57,6 @@ library
   other-modules:
     Lang.Crucible.LLVM.MemModel.Common
     Lang.Crucible.LLVM.MemModel.Generic
-    Lang.Crucible.LLVM.MemModel.MemImpl
     Lang.Crucible.LLVM.MemModel.Value
     Lang.Crucible.LLVM.MemModel.Pointer
     Lang.Crucible.LLVM.MemModel.Type
