+++ crucible-llvm/src/Lang/Crucible/LLVM/MemModel.hs
@@ -34,6 +34,15 @@ module Lang.Crucible.LLVM.MemModel
     Mem
   , memRepr
   , mkMemVar
+  , MemImpl(..)
+  , SomePointer(..)
+  , GlobalMap
+  , emptyMem
+  , memEndian
+  , ppMem
+  , doDumpMem
+  , BlockSource(..)
+  , nextBlock
 
     -- * Pointers
   , LLVMPointerType
@@ -124,9 +133,6 @@ module Lang.Crucible.LLVM.MemModel
   , HasPtrWidth
   , pattern PtrWidth
   , withPtrWidth
-
-    -- * Re-exports
-  , module Lang.Crucible.LLVM.MemModel.MemImpl
   ) where
 
 import           Control.Lens hiding (Empty, (:>))
@@ -137,7 +143,10 @@ import           Control.Monad.Trans (lift)
 import           Control.Monad.Trans.State
 import           Data.Dynamic
 import qualified Data.Map as Map
+import           Data.Map (Map)
+import           Data.IORef
 import           Data.Word
+import           System.IO (Handle, hPutStrLn)
 import           Text.PrettyPrint.ANSI.Leijen hiding ((<$>))
 import           GHC.TypeLits
 
@@ -170,7 +179,6 @@ import           Lang.Crucible.LLVM.MemType
 import           Lang.Crucible.LLVM.MemModel.Type
 import qualified Lang.Crucible.LLVM.MemModel.Generic as G
 import           Lang.Crucible.LLVM.MemModel.Pointer
-import           Lang.Crucible.LLVM.MemModel.MemImpl
 import           Lang.Crucible.LLVM.MemModel.Value
 import           Lang.Crucible.LLVM.Translation.Constant
 import           Lang.Crucible.LLVM.Types
@@ -178,6 +186,48 @@ import           Lang.Crucible.Panic (panic)
 
 import           GHC.Stack
 
+----------------------------------------------------------------------
+-- The MemImpl type
+--
+
+-- | A pointer with an existentially-quantified width
+data SomePointer sym = forall w. SomePointer !(LLVMPtr sym w)
+
+newtype BlockSource = BlockSource (IORef Integer)
+type GlobalMap sym = Map L.Symbol (SomePointer sym)
+
+nextBlock :: BlockSource -> IO Integer
+nextBlock (BlockSource ref) =
+  atomicModifyIORef' ref (\n -> (n+1, n))
+
+-- | The implementation of an LLVM memory, containing an
+-- allocation-block source, global map, handle map, and heap.
+data MemImpl sym =
+  MemImpl
+  { memImplBlockSource :: BlockSource
+  , memImplGlobalMap   :: GlobalMap sym
+  , memImplHandleMap   :: Map Integer Dynamic
+  , memImplHeap        :: G.Mem sym
+  }
+
+memEndian :: MemImpl sym -> EndianForm
+memEndian = G.memEndian . memImplHeap
+
+-- | Produce a fresh empty memory.
+--   NB, we start counting allocation blocks at '1'.
+--   Block number 0 is reserved for representing raw bitvectors.
+emptyMem :: EndianForm -> IO (MemImpl sym)
+emptyMem endianness = do
+  blkRef <- newIORef 1
+  return $ MemImpl (BlockSource blkRef) Map.empty Map.empty (G.emptyMem endianness)
+
+ppMem :: IsExprBuilder sym => MemImpl sym -> Doc
+ppMem mem = G.ppMem (memImplHeap mem)
+
+-- | Pretty print a memory state to the given handle.
+doDumpMem :: IsExprBuilder sym => Handle -> MemImpl sym -> IO ()
+doDumpMem h mem = do
+  hPutStrLn h (show (ppMem mem))
 
 ----------------------------------------------------------------------
 -- Memory operations
